refac: separate frontend pages and components
