+++ frontend/decrypt.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="pages/decrypted.tsx"></script>
+    <script type="module" src="pages/render/decrypted.tsx"></script>
   </body>
 </html>
