+++ frontend/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="pages/index.tsx"></script>
+    <script type="module" src="pages/render/index.tsx"></script>
   </body>
 </html>
