+++ frontend/pages/render/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom/client"
 import React from "react"
 import { HeroUIProvider } from "@heroui/react"
-import { PasteBin } from "../components/PasteBin.js"
+import { PasteBin } from "../PasteBin.js"
 
 const root = ReactDOM.createRoot(document.getElementById("root")!)
